Create SampleStackDriverclass
Input
First Line: contains the total floors N in the building
Second Line onwards: Contains N inputs in which the ith integer denotes the size of the floor
that will be given to architect by factories on the ith day.
Output
You are required to print N lines. Print the size of the floor in descending order, which can be
assembled.
If no floor can be assembled on the ith day, leave the line empty.